+++ frontend/src/app/providers/index.tsx
@@ -2,7 +2,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { AuthProvider } from '@/app/providers/AuthContext'
 import { Router } from '@/app/router'
 
-const queryClient = new QueryClient()
+export const queryClient = new QueryClient()
 
 export function Providers() {
   return (